+++ src/main/java/br/com/zalf/prolog/webservice/Injection.java
@@ -90,6 +90,8 @@ import br.com.zalf.prolog.webservice.frota.socorrorota.SocorroRotaDao;
 import br.com.zalf.prolog.webservice.frota.socorrorota.SocorroRotaDaoImpl;
 import br.com.zalf.prolog.webservice.frota.socorrorota.opcaoproblema.OpcaoProblemaDao;
 import br.com.zalf.prolog.webservice.frota.socorrorota.opcaoproblema.OpcaoProblemaDaoImpl;
+import br.com.zalf.prolog.webservice.frota.socorrorota.relatorio.SocorroRotaRelatorioDao;
+import br.com.zalf.prolog.webservice.frota.socorrorota.relatorio.SocorroRotaRelatorioDaoImp;
 import br.com.zalf.prolog.webservice.frota.veiculo.VeiculoDao;
 import br.com.zalf.prolog.webservice.frota.veiculo.VeiculoDaoImpl;
 import br.com.zalf.prolog.webservice.frota.veiculo.error.VeiculoExceptionHandler;
@@ -189,6 +191,10 @@ public final class Injection {
     @NotNull
     public static SocorroRotaDao provideSocorroDao() {
         return new SocorroRotaDaoImpl();
+    }
+
+    @NotNull
+    public static SocorroRotaRelatorioDao provideSocorroRotaRelatorioDao() { return new SocorroRotaRelatorioDaoImp();
     }
 
     @NotNull